Upgrade to Lucene 4.1
* Removed CustmoMemoryIndex in favor of MemoryIndex which as of 4.1 supports adding the same field twice
* Replaced duplicated logic in X[*]FSDirectory for rate limiting with a RateLimitedFSDirectory wrapper
* Remove hacks to find out merge context in rate limiting in favor of IOContext
* replaced Scorer#freq() return type (from float to int)
* Upgraded FVHighlighter to new 'centered' highlighting
* Fixed RobinEngine to use seperate setCommitData

+++ src/main/java/org/elasticsearch/index/search/nested/BlockJoinQuery.java
@@ -212,7 +212,7 @@ public class BlockJoinQuery extends Query {
         private int parentDoc = -1;
         private int prevParentDoc;
         private float parentScore;
-        private float parentFreq;
+        private int parentFreq;
         private int nextChildDoc;
 
         private int[] pendingChildDocs = new int[5];
@@ -294,10 +294,10 @@ public class BlockJoinQuery extends Query {
                     continue;
                 }
 
-                float totalScore = 0;
-                float totalFreq = 0;
+                int totalScore = 0;
+                int totalFreq = 0;
                 float maxScore = Float.NEGATIVE_INFINITY;
-                float maxFreq = 0;
+                int maxFreq = 0;
 
                 childDocUpto = 0;
                 do {
@@ -313,7 +313,7 @@ public class BlockJoinQuery extends Query {
                     if (scoreMode != ScoreMode.None) {
                         // TODO: specialize this into dedicated classes per-scoreMode
                         final float childScore = childScorer.score();
-                        final float childFreq = childScorer.freq();
+                        final int childFreq = childScorer.freq();
                         pendingChildScores[childDocUpto] = childScore;
                         maxScore = Math.max(childScore, maxScore);
                         maxFreq = Math.max(childFreq, maxFreq);
@@ -364,7 +364,7 @@ public class BlockJoinQuery extends Query {
         }
 
         @Override
-        public float freq() throws IOException {
+        public int freq() throws IOException {
             return parentFreq;
         }
